backend/tgi-scraper/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
black==25.1.0
boto3==1.40.51
botocore==1.40.51
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
Telethon==1.41.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.5.0
uvicorn==0.34.2
